fix(S11P21A705-613): QA 에러 수정

+++ backend/src/main/java/com/zzf/backend/domain/animal/dto/AnimalInfoResponse.java
@@ -13,7 +13,7 @@ public record AnimalInfoResponse(String animalName,
                                  Long totalAssets,
                                  Long totalDeposit,
                                  Long totalSavings,
-                                 Long totalStock,
+                                 Double totalStock,
                                  Long totalLoan,
                                  Long totalCapital) {
 }
